Created dummy home page, moved fab to it for now
